(feat)adds router, refactored search components into searchView
