+++ client/index.html
@@ -20,7 +20,6 @@
 </head>
 <body>
   <div class="col-md-8 col-md-offset-2">
-    <h1>Stack Match</h1>
     <div id="app"></div>
   </div>
   <!-- Compiled code from bundle.js -->
