+++ client/src/components/AppContainer.jsx
@@ -10,66 +10,38 @@ var AppConstants = require('../constants/AppConstants');
 var SearchActionCreators = require('../actions/SearchActionCreators');
 var SearchStore = require('../stores/SearchStore');
 
-var ReactRouter = require('react-router');
-var Router = ReactRouter.Router;
-var Route = ReactRouter.Route;
-var Link = ReactRouter.Link;
+var SearchView = require('./SearchView');
 
-var getStateFromStores = function() {
-  return {
-    results: SearchStore.get()
-  }
-};
+var Router = require('react-router');
+var RouteHandler = Router.RouteHandler;
+var DefaultRoute = Router.DefaultRoute;
+var Route = Router.Route;
+var Link = Router.Link;
+
 
 var AppContainer = React.createClass({
-  // set initial search results to an empty array
-  getInitialState: function() {
-    return {
-      searchResults: []
-    }
-  },
-
-  getSearchStoreState: function() {
-    console.log(SearchStore.get());
-    return SearchStore.get();
-  },
-
-  // Add change listeners
-  componentDidMount: function() {
-    SearchStore.addChangeListener(this._onChange);
-  },
-
-  // Remove change listeners
-  componentWillUnmount: function() {
-    SearchStore.removeChangeListener(this._onChange);
-  },
 
   render: function() {
     return (
       <div>
-        <NavBar />
-        <MainSearchBar />
-        <ResultList list={this.state.searchResults} />
+        <h1>Stack Match</h1>
+        <RouteHandler />
       </div>
     );
-  },
-
-  // Update state when store changes - triggers re-render
-  _onChange: function() {
-    this.setState({searchResults: this.getSearchStoreState()});
   }
+
 });
+
 
 var routes = (
   <Route name='app' path='/' handler={AppContainer}>
-    // <Route name='search' handler={}>
-    // <Route name='trending' handler={}>
-    // <Route name='profile' handler={}>
-    // <Route name='login' handler={}>
-    <DefaultRoute handler={AppContainer}>
+    <Route name='searchView' handler={SearchView}/>
+    <DefaultRoute name='default' handler={SearchView} />
   </Route>
 );
 
-React.render(<AppContainer />, document.getElementById('app'));
+Router.run(routes, function(Handler){
+  React.render(<Handler />, document.getElementById('app'));
+});
 
 module.exports = AppContainer;
