+++ client/src/components/NavButton.jsx
@@ -1,8 +1,11 @@
+var Router = require('react-router');
+var Link = Router.Link;
+
 var NavButton = React.createClass({
   render: function() {
     return (
       <li role="presentation">
-        <a href={this.props.url}>{this.props.label}</a>
+        <Link to={this.props.navTo}>{this.props.label}</Link>
       </li>
     );
   }
