--- client/src/components/UserForm.jsx
@@ -1,8 +1,0 @@
-var UserForm = React.createClass({
-  render: function() {
-    return(
-      <input type='text'>
-      <input type='password'>
-    );
-  }
-});
